Fix a bug in pcmbuf.c when doing offset with modulus.
Causes the track change to go in the wrong buffer (and even be missed
by playback) if the current descriptor for the write index is descriptor
0 because the offset used is "-1" upon track change.

Got nailed by implicit conversion of the % operator dividend to unsigned.

+++ apps/pcmbuf.c
@@ -241,7 +241,7 @@ static struct chunkdesc * index_chunkdesc_offs(size_t index, int offset)
 
     if (offset != 0)
     {
-        i = (i + offset) % pcmbuf_desc_count;
+        i = (i + offset) % (int)pcmbuf_desc_count;
 
         /* remainder => modulus */
         if (i < 0)
